+++ m3_lighthouse/setup_datadir.sh
@@ -1,24 +1,49 @@
 #!/bin/bash
-source ./vars.env
+
+LIGHTHOUSE_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+cd $LIGHTHOUSE_DIR
+
+if [ ! -e ./config.env ]; then
+    echo "did not find ./config.env"
+    exit 1
+fi
+
+source ./config.env
+
+if [ ! -e ../eth2_gen/eth2 ]; then
+    echo "Error: you must generate beacon state and validators before running this script"
+    exit 1
+fi
 
 rm -rf $DATADIR
-mkdir -p $TESTNET_DIR
-mkdir -p $BEACON_DIR/ && cp -ra ./network $BEACON_DIR
+mkdir -p $DATADIR/testnet
+mkdir -p $DATADIR/beacon && cp -ra ./network $DATADIR/beacon/
+mkdir -p $DATADIR/validators
 
-cd $TESTNET_DIR
+if [ ! -e ../eth2_gen/eth2/public ]; then
+    echo "Error: beacon state not generated in ../eth2_gen"
+    exit 1
+fi
+
+cd $DATADIR/testnet
 echo "[]" > ./boot_enr.yaml
-ln -s ../../../eth2_gen/eth2_config.yaml ./config.yaml
+cp ../../../eth2_gen/eth2_config.yaml ./config.yaml
 echo "0" > deploy_block.txt
-ln -s ../../../eth2_gen/eth2/public/genesis.ssz .
+cp ../../../eth2_gen/eth2/public/genesis.ssz .
 
-mkdir -p $VALIDATOR_DIR
-cd $VALIDATOR_DIR
+cd $LIGHTHOUSE_DIR
+if [ ! -e ../eth2_gen/eth2/private ]; then
+    echo "Error: validator keys not generated in ../eth2_gen"
+    exit 1
+fi
+
+cd $DATADIR/validators
 for full in $(echo ../../../eth2_gen/eth2/private/*); do
 	base=$(basename $full);
 	mkdir $base
 	cd $base
-	ln -s ../$full/secrets .
-	ln -s ../$full/keys ./validators
+	cp -r ../$full/secrets .
+	cp -r ../$full/keys ./validators
 	cd ..
 done
